# microoptimization: replace functools.partial with lambda