--- a/mongoengine/fields.py
+++ b/mongoengine/fields.py
@@ -1108,17 +1108,12 @@ class ReferenceField(BaseField):
             collection = self.document_type._get_collection_name()
             value = DBRef(collection, self.document_type.id.to_python(value))
         if isinstance(value, DBRef):
+            fields_copy = _fields[:] if _fields is not None else None
             value = DocumentProxy(
-                functools.partial(
-                    dereference_dbref,
-                    value=value,
-                    document_type=self.document_type,
-                    _lazy_prefetch_base=_lazy_prefetch_base,
-                    _fields=_fields[:] if _fields is not None else None,
-                ),
+                lambda: dereference_dbref(value, self.document_type, _lazy_prefetch_base, _fields),
                 value.id,
                 value.collection,
-                lazy_prefetch_base = _lazy_prefetch_base if _fields else None,
+                lazy_prefetch_base=_lazy_prefetch_base if _fields else None
             )
 
         return value
@@ -1228,14 +1223,9 @@ class CachedReferenceField(BaseField):
                 value = DBRef(collection, self.document_type.id.to_python(value['_id']))
 
         if isinstance(value, DBRef):
+            fields_copy = _fields[:] if _fields is not None else None
             value = DocumentProxy(
-                functools.partial(
-                    dereference_dbref,
-                    value=value,
-                    document_type=self.document_type,
-                    _lazy_prefetch_base=_lazy_prefetch_base,
-                    _fields=_fields[:] if _fields is not None else None,
-                ),
+                lambda: dereference_dbref(value, self.document_type, _lazy_prefetch_base, _fields),
                 value.id,
                 value.collection,
                 lazy_prefetch_base = _lazy_prefetch_base if _fields else None