```
%timeit -n 1000 -r 1000 partial(f, x=1, y=2, z=3)
195 ns ± 57 ns per loop (mean ± std. dev. of 1000 runs, 1000 loops each)

%timeit -n 1000 -r 1000 partial(f, 1, 2, 3)
148 ns ± 12.5 ns per loop (mean ± std. dev. of 1000 runs, 1000 loops each)

%timeit -n 1000 -r 1000 (lambda: f(x=1, y=2, z=3))
68.9 ns ± 31.1 ns per loop (mean ± std. dev. of 1000 runs, 1000 loops each)

%timeit -n 1000 -r 1000 (lambda: f(1, 2, 3))
52.5 ns ± 5.53 ns per loop (mean ± std. dev. of 1000 runs, 1000 loops each)

```
nothing crazy but lambda + using positional args instead of kwargs is a mild win in a hotspot.